refactor: move `AdmonitionInserter` to separate module
